+++ app/App/Account/Account/index.js
@@ -23,7 +23,7 @@ import SignatureRequest from './Requests/SignatureRequest'
 import ChainRequest from './Requests/ChainRequest'
 import AddTokenRequest from './Requests/AddTokenRequest'
 import SignTypedDataRequest from './Requests/SignTypedDataRequest'
-import { isHardwareSigner } from '../../../../resources/utils/signers'
+import { isHardwareSigner } from '../../../../resources/domain/signer'
 
 class _AccountModule extends React.Component {
   // constructor (props, context) {
