--- resources/utils/signers.ts
@@ -1,3 +1,0 @@
-export function isHardwareSigner (type = '') {
-  return ['ledger', 'trezor', 'lattice'].includes(type.toLowerCase())
-}
